web-portal/requirements.txt
# SecureVPN Web Portal Requirements
# Install with: pip install -r requirements.txt

# Core Flask
Flask>=2.3.0
Werkzeug>=2.3.0

# Security
Flask-WTF>=1.2.0  # CSRF protection
WTForms>=3.1.0

# Password hashing
bcrypt>=4.0.0

# QR codes
qrcode[pil]>=7.4.0
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.0

# Optional: Redis for distributed rate limiting
# redis>=5.0.0

# MySQL support
mysql-connector-python>=8.0.0
# OR use SQLAlchemy with PyMySQL:
# SQLAlchemy>=2.0.0
# PyMySQL>=1.1.0
